AutoJagaMAS/requirements.txt
masfactory>=0.1.0
pyyaml>=6.0
loguru>=0.7.0
litellm>=1.0.0
